(PC-37685)[PRO] fix: restore tests against h1 existence within IndividualOfferWizard pages + move getTitle
