+++ pro/src/pages/CollectiveOffer/CollectiveOfferConfirmation/CollectiveOfferConfirmation.module.scss
