+++ pro/src/pages/CollectiveOffer/CollectiveOfferConfirmation/CollectiveOfferConfirmation.tsx
@@ -1,38 +1,206 @@
+import cn from 'classnames'
+
+import { CollectiveOfferDisplayedStatus } from '@/apiClient/v1'
 import { BasicLayout } from '@/app/App/layouts/BasicLayout/BasicLayout'
 import { isCollectiveOfferTemplate } from '@/commons/core/OfferEducational/types'
+import { useActiveFeature } from '@/commons/hooks/useActiveFeature'
 import { RouteLeavingGuardCollectiveOfferCreation } from '@/components/RouteLeavingGuardCollectiveOfferCreation/RouteLeavingGuardCollectiveOfferCreation'
-import { CollectiveOfferConfirmationScreen } from '@/pages/CollectiveOffer/CollectiveOfferConfirmation/CollectiveOfferConfirmation/CollectiveOfferConfirmation'
+import fullValidateIcon from '@/icons/full-validate.svg'
+import fullWaitIcon from '@/icons/full-wait.svg'
+import { ButtonLink } from '@/ui-kit/Button/ButtonLink'
+import { ButtonVariant } from '@/ui-kit/Button/types'
+import { Callout } from '@/ui-kit/Callout/Callout'
+import { SvgIcon } from '@/ui-kit/SvgIcon/SvgIcon'
 
 import {
   type MandatoryCollectiveOfferFromParamsProps,
   withCollectiveOfferFromParams,
 } from '../CollectiveOffer/components/OfferEducational/useCollectiveOfferFromParams'
+import styles from './CollectiveOfferConfirmation.module.scss'
+
+const activeOffer = (institutionDisplayName?: string) => ({
+  title: 'Votre offre a été publiée sur ADAGE',
+  description: (
+    <>
+      Votre offre est désormais visible et réservable par les enseignants et
+      chefs d’établissements{' '}
+      {institutionDisplayName ? (
+        <>
+          de l’établissement scolaire :
+          <br />
+          <br />
+          <b>{institutionDisplayName}</b>
+        </>
+      ) : (
+        'des établissement scolaires qui sont dans la ou les zones de mobilité sélectionnées.'
+      )}
+    </>
+  ),
+  icon: (
+    <SvgIcon
+      src={fullValidateIcon}
+      alt=""
+      className={cn(
+        styles['confirmation-icon'],
+        styles['confirmation-icon-validate']
+      )}
+    />
+  ),
+})
+
+const pendingOffer = {
+  title: 'Offre en cours de validation !',
+  description: (
+    <>
+      Nous vérifions actuellement la conformité de votre offre.
+      <br />
+      <b>Cette vérification pourra prendre jusqu’à 72h.</b>
+      <br />
+      <b>Vous ne pouvez pas effectuer de modification pour l’instant.</b>
+      <br />
+      Vous recevrez un e-mail de confirmation une fois votre offre validée.
+    </>
+  ),
+  icon: (
+    <SvgIcon
+      src={fullWaitIcon}
+      alt=""
+      className={styles['confirmation-icon']}
+    />
+  ),
+}
+
+const showcaseOffer = {
+  title: 'Votre offre a été publiée sur ADAGE',
+  description: (
+    <>
+      Votre offre est visible uniquement par les enseignants et chefs
+      d’établissement de l’Éducation Nationale qui peuvent désormais prendre
+      contact avec vous pour co-construire une offre adaptée à leurs classes.
+    </>
+  ),
+  icon: (
+    <SvgIcon
+      src={fullValidateIcon}
+      alt=""
+      className={cn(
+        styles['confirmation-icon'],
+        styles['confirmation-icon-validate']
+      )}
+    />
+  ),
+}
+
+const getInstitutionDisplayName = (
+  offer: MandatoryCollectiveOfferFromParamsProps['offer']
+) => {
+  if (isCollectiveOfferTemplate(offer)) {
+    return ''
+  }
+
+  if (!offer.institution) {
+    return ''
+  }
+
+  return `${offer.institution.institutionType ?? ''} ${
+    offer.institution.name
+  }`.trim()
+}
+
+const mapOfferStatusToData = (
+  displayedStatus?: CollectiveOfferDisplayedStatus,
+  isShowcase?: boolean,
+  institutionName?: string
+) => {
+  if (displayedStatus === CollectiveOfferDisplayedStatus.UNDER_REVIEW) {
+    return pendingOffer
+  }
+
+  if (isShowcase) {
+    return showcaseOffer
+  }
+
+  return activeOffer(institutionName)
+}
 
 const CollectiveOfferConfirmation = ({
   offer,
 }: MandatoryCollectiveOfferFromParamsProps): JSX.Element => {
-  const getInstitutionDisplayName = () => {
-    if (isCollectiveOfferTemplate(offer)) {
-      return ''
-    }
+  const isShowcase = offer.isTemplate
+  const offerStatus = offer.displayedStatus
+  const offererId = offer.venue.managingOfferer.id
+  const institutionDisplayName = getInstitutionDisplayName(offer)
 
-    if (!offer.institution) {
-      return ''
-    }
-
-    return `${offer.institution.institutionType ?? ''} ${
-      offer.institution.name
-    }`.trim()
-  }
+  const isNewCollectiveOffersStructureActive = useActiveFeature(
+    'WIP_ENABLE_NEW_COLLECTIVE_OFFERS_AND_BOOKINGS_STRUCTURE'
+  )
+  const { title, description, icon } = mapOfferStatusToData(
+    offerStatus,
+    isShowcase,
+    institutionDisplayName
+  )
 
   return (
-    <BasicLayout mainHeading="Votre offre a été publiée sur ADAGE">
-      <CollectiveOfferConfirmationScreen
-        isShowcase={offer.isTemplate}
-        offerStatus={offer.displayedStatus}
-        offererId={offer.venue.managingOfferer.id}
-        institutionDisplayName={getInstitutionDisplayName()}
-      />
+    <BasicLayout mainHeading={title}>
+      <div className={styles['confirmation-wrapper']}>
+        <div className={styles['confirmation']}>
+          {icon}
+          <div>{description}</div>
+          <div className={styles['confirmation-actions']}>
+            <ButtonLink
+              variant={ButtonVariant.SECONDARY}
+              className={styles['confirmation-action']}
+              to={
+                isShowcase && isNewCollectiveOffersStructureActive
+                  ? '/offres/vitrines'
+                  : '/offres/collectives'
+              }
+            >
+              Voir mes offres
+            </ButtonLink>
+            <ButtonLink
+              variant={ButtonVariant.PRIMARY}
+              className={styles['confirmation-action']}
+              to={`/offre/creation${offererId ? `?structure=${offererId}` : ''}`}
+            >
+              Créer une nouvelle offre
+            </ButtonLink>
+          </div>
+        </div>
+
+        <Callout
+          className={styles['confirmation-banner']}
+          title="Quelle est la prochaine étape ?"
+          links={[
+            {
+              href: `https://aide.passculture.app/hc/fr/articles/4416082284945--Acteurs-Culturels-Quel-est-le-cycle-de-vie-de-mon-offre-collective-de-sa-cr%C3%A9ation-%C3%A0-son-remboursement`,
+              label:
+                'Quel est le cycle de vie d’une offre collective, de sa création à son remboursement',
+              isExternal: true,
+            },
+          ]}
+        >
+          {isShowcase ? (
+            <>
+              Les enseignants intéressés par votre offre vitrine vous
+              contacterons par mail ou téléphone. <br />
+              Après un accord mutuel, vous pourrez créer une offre réservable en
+              complétant la date, le prix et l’établissement convenus avec
+              l’enseignant. <br />
+              Cette nouvelle offre apparaitra sur ADAGE et pourra être
+              préréservée par l’enseignant.
+            </>
+          ) : (
+            <>
+              L’enseignant doit préréserver votre offre depuis son compte ADAGE.
+              <br />
+              Une fois la préréservation faite, vous verrez une réservation
+              portant le statut préréservé qui, dans un second temps, devra être
+              officiellement réservée par le chef d’établissement.
+            </>
+          )}
+        </Callout>
+      </div>
       <RouteLeavingGuardCollectiveOfferCreation when={false} />
     </BasicLayout>
   )
