+++ pro/src/pages/IndividualOfferWizard/utils/getTitle.spec.ts
