+++ pro/src/pages/IndividualOfferWizard/utils/getTitle.ts
